const correctness, branched assignment, unused variables, non-standard feature test, example UB, incorrect memset declaration

+++ src/memory.c
@@ -19,6 +19,8 @@
  * FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS
  * IN THE SOFTWARE.
  */
+
+#include <config.h>
 #include <assert.h>
 #include <errno.h>
 #include <fcntl.h>
@@ -52,7 +54,7 @@ struct st_r3_mem_pool_shared_ref_t {
     struct st_r3_mem_pool_shared_entry_t *entry;
 };
 
-void *(*r3_mem__set_secure)(void *, int, unsigned int) = memset;
+void *(*r3_mem__set_secure)(void *, int, size_t) = memset;
 
 static __thread r3_mem_recycle_t mempool_allocator = {16};
